requirements.txt
bertopic
google-api-python-client>=2.190.0
google-auth-oauthlib>=1.2.2
matplotlib
numpy>=1.26,<3
openai
pandas>=2.2,<3
plotly
pytrends
python-dotenv>=1.0,<2
requests>=2.31,<3
scikit-learn
streamlit==1.54.0
altair==5.5.0
tqdm
pytest
streamlit-option-menu>=0.3.12
streamlit-lottie>=0.0.5
yt-dlp>=2026.3.3
youtube-transcript-api>=1.2.4
